Adapt Serializer and SerializationException for STELLA
The serializer #includes the needed headers from STELLA and
defines the correct API for writing and reading into/from a
STELLA data field. The information about the storage directions
and the strides is inferred from the field itself. The interface
to initialize a field has been removed.

The #include to the Serialbox Exception.h header has been remove to
prevent the exposure of C++11 code. The SERIALBOX_NOEXCEPT macro
is defined again here.