Reading into a STELLA field obeys the new format
The format has changed for degenerated dimensions, which are now
stripped. With this patch, these are also stripped in the STELLA
case.

Notice we removed a check for the consistency of the boundaries.
Fixing this check would have required a lot of boilerplate and it
has always proved to be non necessary. It only depends on the
consistency of the STELLA verification framework anyway, which is
consistent now and won't be touched again in the future, so we are on
the safe side anyway.