[Fortran] FortranTestGenerator frontend (#134)
This PR adds the new Fortran module `m_ser_ftg.f90` that contains the interface for the [FortranTestGenerator](https://github.com/fortesg/fortrantestgenerator) (FTG) and the corresponding test `ser_ftg_test.pf`. 

We add this interface for a client application for two reasons: a) This pattern already exists within serialbox2 with the frontends to gridtools and stella. b) Since FTG is pure Python and only generates Fortran, it has no place for this module and having a separate repository for the FTG->SB2 interface would be difficult to maintain.

Contents of `m_ser_ftg.f90`:

- Module variables for the current serializer and savepoint in use.
- Wrapper routines for the routines from `m_serialize.f90` without the serializer and savepoint arguments using the ones from the module variables.
- Routines for allocating arrays based on the stored sizes and bounds (halos).
- Write: Ignoring of not-allocated/associated (pointer) variables when they can be detected.
- Read: Ignoring of non-existing fields.
- A routine for registering fields of arbitrary types without actually writing the data. Since serialbox2 currently don't allow arbitrary types, this is a hack that sets `'int'` as type and the actual type as  metainfo.