compare.py: add support for filtering fields (#250)
Small PR to add the option -f to be able to filter the fields one wants to check with giving a substr as argument. Only fields including the substrings are then checked and not all of them. Allows to also compare one field or some types of fields quickly which speeds up the workflow quite a lot (for me at least).