Fix python build (#264)
Added `${ORIGIN}` to the build `rpath` for python build

when running the python build with pip 
```
cd src/serialbox-python
pip install .
```
the cmake build target is run, after that the shared libraries are copied from the build folder to `site_packages` folder of the python environment. Adding ` ${ORIGIN}` to the `rpath` allows `cdll.LoadLibrary`  to find dependent libraries in the current folder.